+++ source/state_representation/include/state_representation/space/cartesian/CartesianAcceleration.hpp
@@ -49,7 +49,7 @@ public:
   /**
    * @brief Empty constructor
    */
-  explicit CartesianAcceleration() = default;
+  explicit CartesianAcceleration();
 
   /**
    * @brief Constructor with name and reference frame provided