+++ source/state_representation/src/space/joint/JointTorques.cpp
@@ -3,6 +3,10 @@
 using namespace state_representation::exceptions;
 
 namespace state_representation {
+JointTorques::JointTorques() {
+  this->set_type(StateType::JOINT_TORQUES);
+}
+
 JointTorques::JointTorques(const std::string& robot_name, unsigned int nb_joints) : JointState(robot_name, nb_joints) {
   this->set_type(StateType::JOINT_TORQUES);
 }