+++ source/state_representation/src/space/joint/JointVelocities.cpp
@@ -4,6 +4,10 @@
 using namespace state_representation::exceptions;
 
 namespace state_representation {
+JointVelocities::JointVelocities() {
+  this->set_type(StateType::JOINT_VELOCITIES);
+}
+
 JointVelocities::JointVelocities(const std::string& robot_name, unsigned int nb_joints) :
     JointState(robot_name, nb_joints) {
   this->set_type(StateType::JOINT_VELOCITIES);